+++ src/components/omniSidebarNav/allGuides.testfixture.js
@@ -6163,4 +6163,4 @@ const AllGuides = {
   }
 }
 
-export default AllGuides
+export default AllGuides;
