+++ src/components/omniSidebarNav/helpers.js
@@ -1,4 +1,6 @@
 
+import allGuides from './allGuidesQuery';
+
 // Searches through the top level items find the subdata structure that
 // matches the parent path requested.
 // todo, rename "maximumParent" to something better.
@@ -11,4 +13,44 @@ const findSubMenuItemsToUse = function (maximumParent, NestedItems) {
   return undefined;
 }
 
-export {findSubMenuItemsToUse };
+function getChildrenForGuideDirectory(AllTheGuides, guideDirectory) {
+  const ChildItems = [];
+  for (let item of AllTheGuides) {
+    if (item.node.fields.guide_directory === guideDirectory) {
+      ChildItems.push({
+        link: item.node.fields.slug,
+        title: item.node.frontmatter.subtitle,
+      });
+    }
+  }
+  return ChildItems;
+}
+
+
+// @todo, flip these arguments right.
+const getTitleForGuideDirectory = function (guideDirectory, AllTheGuides) {
+  for (let item of AllTheGuides) {
+    if (item.node.fields.slug === '/' + guideDirectory) {
+      return item.node.frontmatter.title;
+    }
+  }
+}
+
+// return an item list for a guide directory.
+// guideDirectory is the directory name of the guide.
+// AllGuides is the result of the allGuides query.
+// For testing purposes, AllGuides can be directly set with the fixture
+// allGuides.testfixtures.js
+const getGuideDirectory = (guideDirectory, AllGuides = allGuides()) => {
+  return {
+    link: '/' + guideDirectory,
+    title: getTitleForGuideDirectory(guideDirectory, AllGuides.allGuides.edges),
+    children: getChildrenForGuideDirectory(
+      AllGuides.allGuides.edges,
+      guideDirectory,
+    ),
+  };
+};
+
+
+export { findSubMenuItemsToUse, getGuideDirectory, getTitleForGuideDirectory };
