+++ src/components/omniSidebarNav/helpers.test.js
@@ -1,6 +1,7 @@
 import { expect, test } from 'vitest'
 import { sum } from './sum'
-import { findSubMenuItemsToUse } from './helpers'
+import { findSubMenuItemsToUse, getTitleForGuideDirectory } from './helpers'
+import AllGuides from './allGuides.testfixture'
 
 test('adds 1 + 2 to equal 3', () => {
   expect(sum(1, 2)).toBe(3)
@@ -22,3 +23,11 @@ test('finds sub menu to use', () => {
   expect(sum(1, 2)).toBe(3)
 })
 
+
+test('Find the guide title by directory', () => {
+  const guideDirectory = "guides/account-mgmt/account";
+
+
+  expect(getTitleForGuideDirectory(guideDirectory, AllGuides.allGuides.edges)).toEqual("Manage Your Account")
+  //expect(sum(1, 2)).toBe(3)
+})
